\name{KANT-package}
\alias{KANT-package}
\alias{KANT}
\docType{package}
\title{
KANT : Kancer Antigene Tracker
}
\description{
This algorithm was designed to identify and sort genes, associated to transmembrane protein, found to be more strongly expressed in a tumor dataset than in a normal dataset.
It can be used in other conditions. 

}
\details{
\tabular{ll}{
Package: \tab KANT\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2014-07-30\cr
License: \tab GPL\cr
}


} 

\author{
Noemie Robil

Maintainer: Who to complain to <noemie.robil@ligue-cancer.net>

}
\references{
KANT : a gene expression-based tool for detecting putative membrane cancer-specific antigens, not published
}



