\name{TM}
\alias{TM}
\docType{data}
\title{
List of geneid linked with transmembrane proteins
}
\description{
GeneID of genes associated with proteins identified as transmembrane by MEMSAT-SVM, among the Uniprot swissprot database from july 2012.
}
\usage{data(TM)}
\format{
  A data frame with 4827 observations on the following variable.

}

\keyword{datasets}
