\name{TM}
\alias{TM}
\docType{data}
\title{
List of geneid linked with transmembrane proteins
}
\description{
GeneID of genes coding for proteins identified as transmembrane by MEMSAT-SVM (modified version), among the Uniprot swissprot database from february 2014.
}
\usage{data(TM)}
\format{
  A data frame with 4964 observations on the following variable.

}


